[chg] 升级工程版本到vs2019。 [chg] 去除编译告警。 [chg] 修改星体数量可设置范围。 [chg] 修改窗体初始透明度。 [chg] 优化部分代码，方便阅读。 [del] 删除临时文件。

+++ StarrySkyAndSnake/StarrySkyAndSnake/Snake.cpp
@@ -29,7 +29,7 @@ CSnake::CSnake()
 	m_UnDead = false;
 	goon = 0;
 	Add(32,32);
-	srand(time(0));
+	srand((unsigned int)time(0));
 	apple.m=(rand()%30)*16;
 	apple.n=(rand()%30)*16;
 }
@@ -218,7 +218,7 @@ void CSnake::GameOver()
 		direction = 4;
 		m_list.RemoveAll();
 		Add(32,32);
-		srand(time(0));
+		srand((unsigned int)time(0));
 		apple.m=(rand()%10)*16;
 		apple.n=(rand()%10)*16;
 		glo=0;
